+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Show progressbar of items being sent by danbooru search / latest
 - Add @XenianBot to danbooru search / latest results
+- Remember where danbooru files where downloaded too
 
 
 2.3.0 (2019-01-05)